services/shared-python-lib/pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "home-service-shared"
version = "0.1.0"
dependencies = [
    "sqlalchemy",
    "psycopg2-binary",
    "python-dotenv",
    "pydantic-settings>=2.0",
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-exporter-otlp",
    "opentelemetry-instrumentation-fastapi",
    "opentelemetry-instrumentation-sqlalchemy",
    "opentelemetry-instrumentation-logging",
    "opentelemetry-instrumentation-requests",
]
